+++ wijisqs/broker.py
@@ -11,7 +11,6 @@ import concurrent
 import botocore.config
 import botocore.session
 
-from . import utils
 from . import buffer
 
 if typing.TYPE_CHECKING:
@@ -88,7 +87,6 @@ class SqsBroker(wiji.broker.BaseBroker):
             )
         self.DelaySeconds = DelaySeconds
 
-        self.QueueUrl: typing.Union[None, str] = None
         self.aws_region_name = aws_region_name
         self.aws_access_key_id = aws_access_key_id
         self.aws_secret_access_key = aws_secret_access_key
@@ -108,6 +106,8 @@ class SqsBroker(wiji.broker.BaseBroker):
 
         self.task_receipt: typing.Dict[str, str] = {}
         self._thread_name_prefix: str = "wiji-SqsBroker-thread-pool"
+
+        self.queue_name_and_url: typing.Dict[str, str] = {}
 
         self.recieveBuf: buffer.ReceiveBuffer = buffer.ReceiveBuffer()
         self.sendBuf: buffer.SendBuffer = buffer.SendBuffer()
@@ -368,8 +368,13 @@ class SqsBroker(wiji.broker.BaseBroker):
                 executor, functools.partial(self._tag_queue, queue_name=queue_name)
             )
 
-    @utils.execute_only_once
     def _create_queue(self, queue_name: str) -> None:
+        """
+        this needs to run for as many queue_name's as will be handled by this broker
+        """
+        if self.queue_name_and_url.get(queue_name):
+            # already exists
+            return
         try:
             response = self.client.create_queue(
                 QueueName=queue_name,
@@ -385,25 +390,33 @@ class SqsBroker(wiji.broker.BaseBroker):
             )
             response.update({"event": "wijisqs.SqsBroker._create_queue", "queue_name": queue_name})
             self.logger.log(logging.DEBUG, response)
-            self.QueueUrl = response["QueueUrl"]
-            setattr(self, "QueueUrl", response["QueueUrl"])
+            self.queue_name_and_url[queue_name] = response["QueueUrl"]
         except Exception as e:
             raise e
 
-    @utils.execute_only_once
     def _tag_queue(self, queue_name: str) -> None:
-        response = self.client.tag_queue(QueueUrl=self.QueueUrl, Tags=self.queue_tags)
+        """
+        this needs to run for as many queue_name's as will be handled by this broker
+        """
+        response = self.client.tag_queue(
+            QueueUrl=self.queue_name_and_url[queue_name], Tags=self.queue_tags
+        )
         response.update({"event": "wijisqs.SqsBroker._tag_queue", "queue_name": queue_name})
         self.logger.log(logging.DEBUG, response)
 
     def _get_queue_url(self, queue_name: str) -> None:
-        # this should always run during `check` call
+        """
+        this should always run during `check` call
+        it populates the queue_name and QueueUrl mapping
+        """
+        if self.queue_name_and_url.get(queue_name):
+            # already exists
+            return
         try:
             response = self.client.get_queue_url(QueueName=queue_name)
             response.update({"event": "wijisqs.SqsBroker._get_queue_url", "queue_name": queue_name})
             self.logger.log(logging.DEBUG, response)
-            self.QueueUrl = response["QueueUrl"]
-            setattr(self, "QueueUrl", response["QueueUrl"])
+            self.queue_name_and_url[queue_name] = response["QueueUrl"]
         except Exception as e:
             raise e
 
@@ -501,7 +514,7 @@ class SqsBroker(wiji.broker.BaseBroker):
     ) -> None:
         delay = self._calculate_msg_delay(task_options=task_options)
         response = self.client.send_message(
-            QueueUrl=self.QueueUrl,
+            QueueUrl=self.queue_name_and_url[queue_name],
             MessageBody=item,
             # DelaySeconds is the length of time, in seconds, for which to delay a specific message
             DelaySeconds=delay,
@@ -525,7 +538,9 @@ class SqsBroker(wiji.broker.BaseBroker):
         # TODO: validate size
         # The maximum allowed individual message size and the maximum total payload size
         # (the sum of the individual lengths of all of the batched messages) are both 256 KB (262,144 bytes).
-        response = self.client.send_message_batch(QueueUrl=self.QueueUrl, Entries=Entries)
+        response = self.client.send_message_batch(
+            QueueUrl=self.queue_name_and_url[queue_name], Entries=Entries
+        )
         response.update(
             {"event": "wijisqs.SqsBroker._send_message_batch", "queue_name": queue_name}
         )
@@ -590,7 +605,7 @@ class SqsBroker(wiji.broker.BaseBroker):
         1. https://botocore.amazonaws.com/v1/documentation/api/latest/reference/services/sqs.html#SQS.Client.receive_message
         """
         response = self.client.receive_message(
-            QueueUrl=self.QueueUrl,
+            QueueUrl=self.queue_name_and_url[queue_name],
             AttributeNames=["All"],
             MessageAttributeNames=["All"],
             MaxNumberOfMessages=self.MaxNumberOfMessages,
@@ -670,7 +685,7 @@ class SqsBroker(wiji.broker.BaseBroker):
         ReceiptHandle = self.task_receipt.pop(task_options.task_id, None)
         if ReceiptHandle:
             response = self.client.delete_message(
-                QueueUrl=self.QueueUrl, ReceiptHandle=ReceiptHandle
+                QueueUrl=self.queue_name_and_url[queue_name], ReceiptHandle=ReceiptHandle
             )
             response.update(
                 {"event": "wijisqs.SqsBroker._delete_message", "queue_name": queue_name}