+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## `wijisqs` changelog:
 most recent version is listed first.
 
+## **version:** v0.1.8
+- one `wijisqs.SqsBroker` instance should be able to serve multiple queues: https://github.com/komuw/wijisqs/pull/26
 
 ## **version:** v0.1.6
 - bugfix, make sure QueueUrl is always available: https://github.com/komuw/wijisqs/pull/23